+++ tests/testthat/test-tsibble.R
@@ -42,7 +42,7 @@ df <- data.frame(time = x, value = rnorm(10))
 tsbl <- as_tsibble(df, index = time)
 
 test_that("POSIXct with 10 milliseconds interval", {
-  expect_output(print(tsbl), "A tsibble: 10 x 2 \\[0.01SECOND\\]")
+  expect_output(print(tsbl), "A tsibble: 10 x 2 \\[10MILLI\\]")
 })
 
 library(nanotime)
@@ -51,7 +51,7 @@ df <- data.frame(time = x, value = rnorm(10))
 tsbl <- as_tsibble(df)
 
 test_that("nanotime with 1 nanoseconds interval", {
-  expect_output(print(tsbl), "A tsibble: 10 x 2 \\[1e-09SECOND\\]")
+  expect_output(print(tsbl), "A tsibble: 10 x 2 \\[1NANO\\]")
 })
 
 test_that("POSIXt with 1 second interval", {